refactor: add VideoComponent
