+++ frontend/src/App.tsx
@@ -1,17 +1,10 @@
-import React, { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
+import { VideoInterface } from "./components/Video/VideoInterface";
+import VideoComponent from "./components/Video/VideoComponent";
 import "./App.css";
 
-interface VideoProps {
-  id: number;
-  title: string;
-  url: string;
-  thumbnailUrl: string;
-  publish_date: string;
-  todays: boolean;
-}
-
 function App() {
-  const [video, setVideo] = useState<VideoProps>();
+  const [video, setVideo] = useState<VideoInterface>();
 
   const getVideo = useCallback(async () => {
     const response = await fetch(
@@ -26,22 +19,7 @@ function App() {
   }, [getVideo]);
 
   return (
-    <div className="App">
-      {video && (
-        <div>
-          <iframe
-            width="200"
-            height="200"
-            src={video.url.replace("watch?v=", "embed/")}
-            title="YouTube video player"
-            frameBorder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowFullScreen
-          ></iframe>
-          <h1 className="video-title">{video.title}</h1>
-        </div>
-      )}
-    </div>
+    <div className="App">{video && <VideoComponent url={video.url} />}</div>
   );
 }
